finish GHz and MHz distinctions, finish update interval selection

+++ pycpu2.py
@@ -9,7 +9,7 @@ command = input("Choose a function:\n1. CPU Frequency (MHz)\n2. CPU Temperature 
 
 if command == "1":
     unit = input("Choose unit:\n1.MHz\n2.GHz\n")
-    interval = int(input("Choose update interval (in seconds):\n"))
+    interval = float(input("Choose update interval (in seconds):\n"))
     if unit == "2":
         while True:
             print(str(round(psutil.cpu_freq()[0], -1) / 1000) + " GHz")
